fix(types): export `NuxtVueI18n` namespace to allow to annotate configuration
Also moved nuxt specific types to new file and leave `vue.d.ts` for
types that extend vue and nuxt.
