+++ types/index.d.ts
@@ -1,2 +1,4 @@
 // augment typings of Vue.js
-import "./vue";
+import './vue'
+
+export { NuxtVueI18n } from './nuxt-i18n'
